wip more rershuffling
