+++ server/instance/instance_cloud.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License for license information.
 
-package main
+package instance
 
 import (
 	"context"
