+++ server/instance/instance_server.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License for license information.
 
-package main
+package instance
 
 import (
 	"crypto/x509"
