+++ server/jira/command.go
