+++ server/store/kv_mock_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-present Mattermost, Inc. All Rights Reserved.
 // See License for license information.
 
-package main
+package store
 
 import (
 	"github.com/andygrunwald/go-jira"
